Revert "internal-0007-PATCH-1-1-avformat-dashdec-prevent-duration-wrap.patch"
This reverts commit 1f95cf8d48b021eb4d435163d4427aea14a1a554.

+++ libavformat/dashdec.c
@@ -209,12 +209,12 @@ static uint64_t get_utc_date_time_insec(AVFormatContext *s, const char *datetime
     return av_timegm(&timeinfo);
 }
 
-static uint64_t get_duration(AVFormatContext *s, const char *duration, int inMicroseconds)
+static uint32_t get_duration(AVFormatContext *s, const char *duration, int inMicroseconds)
 {
     /* ISO-8601 duration parser */
-    uint64_t days = 0;
-    uint64_t hours = 0;
-    uint64_t mins = 0;
+    uint32_t days = 0;
+    uint32_t hours = 0;
+    uint32_t mins = 0;
     float secs = 0;
     int size = 0;
     float value = 0;
@@ -233,13 +233,13 @@ static uint64_t get_duration(AVFormatContext *s, const char *duration, int inMic
         }
         switch (type) {
         case 'D':
-            days = (uint64_t)value;
+            days = (uint32_t)value;
             break;
         case 'H':
-            hours = (uint64_t)value;
+            hours = (uint32_t)value;
             break;
         case 'M':
-            mins = (uint64_t)value;
+            mins = (uint32_t)value;
             break;
         case 'S':
             secs = value;
@@ -251,9 +251,9 @@ static uint64_t get_duration(AVFormatContext *s, const char *duration, int inMic
         ptr += size;
     }
     if (inMicroseconds) {
-        return  (((days * 24 + hours) * 60 + mins) * 60 * AV_TIME_BASE) + (uint64_t)(secs * AV_TIME_BASE);
+        return  (((days * 24 + hours) * 60 + mins) * 60 * AV_TIME_BASE) + (uint32_t)(secs * AV_TIME_BASE);
     }
-    return  ((days * 24 + hours) * 60 + mins) * 60 + (uint64_t)secs;
+    return  ((days * 24 + hours) * 60 + mins) * 60 + (uint32_t)secs;
 }
 
 static uint32_t get_duration_inmicrosec(AVFormatContext *s, const char *duration)
@@ -1258,8 +1258,8 @@ static int parse_manifest(AVFormatContext *s, const char *url, AVIOContext *in)
     xmlNodePtr adaptionset_node = NULL;
     xmlAttrPtr attr = NULL;
     char *val  = NULL;
-    uint64_t period_duration_sec = 0;
-    uint64_t period_start_sec = 0;
+    uint32_t period_duration_sec = 0;
+    uint32_t period_start_sec = 0;
 
     if (!in) {
         close_in = 1;
